+++ sites/m.dishen.com/components/mip-dishen-datepicker/mip-dishen-datepicker.js
@@ -1,5 +1,6 @@
 import './style/index.less'
-
+import calendar from './calendar'
+import DateScroll from './DateScroll'
 class RuiDatepicker {
   constructor (config) {
     this.gearDate = null
@@ -18,100 +19,7 @@ class RuiDatepicker {
   }
   bindEvent (type) {
     let _self = this
-    // 农历数据
-    let LunarCal = [
-      new TagLunarCal(38, 0, 0, 38, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1), /* 1940 */
-      new TagLunarCal(26, 6, 2, 44, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0),
-      new TagLunarCal(45, 0, 3, 49, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0),
-      new TagLunarCal(35, 0, 4, 54, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1),
-      new TagLunarCal(24, 4, 5, 59, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1), /* 1944 */
-      new TagLunarCal(43, 0, 0, 5, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1),
-      new TagLunarCal(32, 0, 1, 10, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1),
-      new TagLunarCal(21, 2, 2, 15, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1),
-      new TagLunarCal(40, 0, 3, 20, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1), /* 1948 */
-      new TagLunarCal(28, 7, 5, 26, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1),
-      new TagLunarCal(47, 0, 6, 31, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1),
-      new TagLunarCal(36, 0, 0, 36, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0),
-      new TagLunarCal(26, 5, 1, 41, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1), /* 1952 */
-      new TagLunarCal(44, 0, 3, 47, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1),
-      new TagLunarCal(33, 0, 4, 52, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0),
-      new TagLunarCal(23, 3, 5, 57, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1),
-      new TagLunarCal(42, 0, 6, 2, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1), /* 1956 */
-      new TagLunarCal(30, 8, 1, 8, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0),
-      new TagLunarCal(48, 0, 2, 13, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 0),
-      new TagLunarCal(38, 0, 3, 18, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1),
-      new TagLunarCal(27, 6, 4, 23, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0), /* 1960 */
-      new TagLunarCal(45, 0, 6, 29, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0),
-      new TagLunarCal(35, 0, 0, 34, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1),
-      new TagLunarCal(24, 4, 1, 39, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0),
-      new TagLunarCal(43, 0, 2, 44, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0), /* 1964 */
-      new TagLunarCal(32, 0, 4, 50, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1),
-      new TagLunarCal(20, 3, 5, 55, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0),
-      new TagLunarCal(39, 0, 6, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 0, 1, 0),
-      new TagLunarCal(29, 7, 0, 5, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1), /* 1968 */
-      new TagLunarCal(47, 0, 2, 11, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1),
-      new TagLunarCal(36, 0, 3, 16, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0),
-      new TagLunarCal(26, 5, 4, 21, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0, 1),
-      new TagLunarCal(45, 0, 5, 26, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1), /* 1972 */
-      new TagLunarCal(33, 0, 0, 32, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1),
-      new TagLunarCal(22, 4, 1, 37, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1),
-      new TagLunarCal(41, 0, 2, 42, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1),
-      new TagLunarCal(30, 8, 3, 47, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1), /* 1976 */
-      new TagLunarCal(48, 0, 5, 53, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1),
-      new TagLunarCal(37, 0, 6, 58, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1),
-      new TagLunarCal(27, 6, 0, 3, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0),
-      new TagLunarCal(46, 0, 1, 8, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0), /* 1980 */
-      new TagLunarCal(35, 0, 3, 14, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1),
-      new TagLunarCal(24, 4, 4, 19, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1),
-      new TagLunarCal(43, 0, 5, 24, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1),
-      new TagLunarCal(32, 10, 6, 29, 1, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1), /* 1984 */
-      new TagLunarCal(50, 0, 1, 35, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0),
-      new TagLunarCal(39, 0, 2, 40, 0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1),
-      new TagLunarCal(28, 6, 3, 45, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1, 0, 0),
-      new TagLunarCal(47, 0, 4, 50, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1), /* 1988 */
-      new TagLunarCal(36, 0, 6, 56, 1, 0, 0, 1, 0, 1, 0, 1, 0, 1, 1, 1, 0),
-      new TagLunarCal(26, 5, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 1),
-      new TagLunarCal(45, 0, 1, 6, 0, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0),
-      new TagLunarCal(34, 0, 2, 11, 0, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0), /* 1992 */
-      new TagLunarCal(22, 3, 4, 17, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0),
-      new TagLunarCal(40, 0, 5, 22, 1, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0),
-      new TagLunarCal(30, 8, 6, 27, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1, 0, 1),
-      new TagLunarCal(49, 0, 0, 32, 0, 1, 0, 1, 1, 0, 1, 0, 1, 1, 0, 0, 1), /* 1996 */
-      new TagLunarCal(37, 0, 2, 38, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1),
-      new TagLunarCal(27, 5, 3, 43, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 0, 1),
-      new TagLunarCal(46, 0, 4, 48, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1), /* 1999 */
-      new TagLunarCal(35, 0, 5, 53, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1), /* 2000 */
-      new TagLunarCal(23, 4, 0, 59, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1),
-      new TagLunarCal(42, 0, 1, 4, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1),
-      new TagLunarCal(31, 0, 2, 9, 1, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0),
-      new TagLunarCal(21, 2, 3, 14, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1), /* 2004 */
-      new TagLunarCal(39, 0, 5, 20, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1),
-      new TagLunarCal(28, 7, 6, 25, 1, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1),
-      new TagLunarCal(48, 0, 0, 30, 0, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 1),
-      new TagLunarCal(37, 0, 1, 35, 1, 0, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1), /* 2008 */
-      new TagLunarCal(25, 5, 3, 41, 1, 1, 0, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1),
-      new TagLunarCal(44, 0, 4, 46, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1),
-      new TagLunarCal(33, 0, 5, 51, 1, 0, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1),
-      new TagLunarCal(22, 4, 6, 56, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0), /* 2012 */
-      new TagLunarCal(40, 0, 1, 2, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0),
-      new TagLunarCal(30, 9, 2, 7, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1),
-      new TagLunarCal(49, 0, 3, 12, 0, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 1),
-      new TagLunarCal(38, 0, 4, 17, 1, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0), /* 2016 */
-      new TagLunarCal(27, 6, 6, 23, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1),
-      new TagLunarCal(46, 0, 0, 28, 0, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 0),
-      new TagLunarCal(35, 0, 1, 33, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0),
-      new TagLunarCal(24, 4, 2, 38, 0, 1, 1, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1), /* 2020 */
-      new TagLunarCal(42, 0, 4, 44, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1),
-      new TagLunarCal(31, 0, 5, 49, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0),
-      new TagLunarCal(21, 2, 6, 54, 0, 1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 0, 1),
-      new TagLunarCal(40, 0, 0, 59, 0, 1, 0, 0, 1, 0, 1, 1, 0, 1, 1, 0, 1), /* 2024 */
-      new TagLunarCal(28, 6, 2, 5, 1, 0, 1, 0, 0, 1, 0, 1, 0, 1, 1, 1, 0),
-      new TagLunarCal(47, 0, 3, 10, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 1, 0, 1),
-      new TagLunarCal(36, 0, 4, 15, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0, 1),
-      new TagLunarCal(25, 5, 5, 20, 1, 1, 1, 0, 1, 0, 0, 1, 0, 0, 1, 1, 0), /* 2028 */
-      new TagLunarCal(43, 0, 0, 26, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0, 1, 0, 1),
-      new TagLunarCal(32, 0, 1, 31, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 1, 0, 0)
-    ]
+
     // 初始化年月日默认值
     function initTriger () {
       if (_self.trigger.getAttribute('data-date')) {
@@ -137,9 +45,8 @@ class RuiDatepicker {
           let dateYY = dateArr.yy
           let dateMM = dateArr.mm
           let dateDD = dateArr.dd
-          let objDate = calendarConvert(0, dateYY, dateMM, dateDD)// 返回转换对象
-          // let isRm = LunarCal[objDate.yy - _self.minY].Intercalation ? LunarCal[objDate.yy - _self.minY].Intercalation : 0// 判断是不是闰年
-          _self.trigger.value = objDate.yy + '年' + getChinese('mm', objDate.mm) + getChinese('dd', objDate.dd) + (hourVal === -1 ? '' : ' ' + getChinese('hh', hourVal) + '时')
+          let objDate = calendar.calendarConvert(0, dateYY, dateMM, dateDD)// 返回转换对象
+          _self.trigger.value = objDate.yy + '年' + calendar.getChinese('mm', objDate.mm) + calendar.getChinese('dd', objDate.dd) + (hourVal === -1 ? '' : ' ' + calendar.getChinese('hh', hourVal) + '时')
         } else {
           _self.trigger.value = dateArr.yy + '年' + dateArr.mm + '月' + dateArr.dd + '日' + (hourVal === -1 ? '' : ' ' + hourVal + '时')
         }
@@ -237,44 +144,14 @@ class RuiDatepicker {
       let dateDD = _self.gearDate.querySelector('.dateDD')
       let dateHH = _self.gearDate.querySelector('.dateHH')
       let dateMin = _self.gearDate.querySelector('.dateMin')
-      if (dateYY) { dateYY.addEventListener('touchstart', gearTouchStart) }
-      if (dateMM) { dateMM.addEventListener('touchstart', gearTouchStart) }
-      if (dateDD) { dateDD.addEventListener('touchstart', gearTouchStart) }
-      if (dateHH) { dateHH.addEventListener('touchstart', gearTouchStart) }
-      if (dateMin) { dateMin.addEventListener('touchstart', gearTouchStart) }
-      if (dateYY) { dateYY.addEventListener('mousedown', gearMouseSlide) }
-      if (dateMM) { dateMM.addEventListener('mousedown', gearMouseSlide) }
-      if (dateDD) { dateDD.addEventListener('mousedown', gearMouseSlide) }
-      if (dateHH) { dateHH.addEventListener('mousedown', gearMouseSlide) }
-      if (dateMin) { dateMin.addEventListener('mousedown', gearMouseSlide) }
-      if (dateYY) { dateYY.addEventListener('touchmove', gearTouchMove) }
-      if (dateMM) { dateMM.addEventListener('touchmove', gearTouchMove) }
-      if (dateDD) { dateDD.addEventListener('touchmove', gearTouchMove) }
-      if (dateHH) { dateHH.addEventListener('touchmove', gearTouchMove) }
-      if (dateMin) { dateMin.addEventListener('touchmove', gearTouchMove) }
-      if (dateYY) { dateYY.addEventListener('touchend', gearTouchEnd) }
-      if (dateMM) { dateMM.addEventListener('touchend', gearTouchEnd) }
-      if (dateDD) { dateDD.addEventListener('touchend', gearTouchEnd) }
-      if (dateHH) { dateHH.addEventListener('touchend', gearTouchEnd) }
-      if (dateMin) { dateMin.addEventListener('touchend', gearTouchEnd) }
+
+      if (dateYY) { _self.yyScroll = new DateScroll(dateYY.getAttribute('val')).init(dateYY) }
+      if (dateMM) { _self.mmScroll = new DateScroll(dateMM.getAttribute('val')).init(dateMM) }
+      if (dateDD) { _self.ddScroll = new DateScroll(dateDD.getAttribute('val')).init(dateDD) }
+      if (dateHH) { _self.hhScroll = new DateScroll(dateHH.getAttribute('val')).init(dateHH) }
+      if (dateMin) { _self.minScroll = new DateScroll(dateMin.getAttribute('val')).init(dateMin) }
       // 阻止透视滑动
       _self.gearDate.addEventListener('touchmove', function (e) { e.preventDefault() })
-      // 阻止鼠标滚轮事件
-      if (navigator.userAgent.indexOf('Firefox') > 0) {
-        _self.gearDate.addEventListener('DOMMouseScroll', function (e) { e.preventDefault() }, false)
-        if (dateYY) { dateYY.addEventListener('DOMMouseScroll', gearMouseRolling, false) }
-        if (dateMM) { dateMM.addEventListener('DOMMouseScroll', gearMouseRolling, false) }
-        if (dateDD) { dateDD.addEventListener('DOMMouseScroll', gearMouseRolling, false) }
-        if (dateHH) { dateHH.addEventListener('DOMMouseScroll', gearMouseRolling, false) }
-        if (dateMin) { dateMin.addEventListener('DOMMouseScroll', gearMouseRolling, false) }
-      } else {
-        _self.gearDate.onmousewheel = function (e) { return false }
-        if (dateYY) { dateYY.onmousewheel = gearMouseRolling }
-        if (dateMM) { dateMM.onmousewheel = gearMouseRolling }
-        if (dateDD) { dateDD.onmousewheel = gearMouseRolling }
-        if (dateHH) { dateHH.onmousewheel = gearMouseRolling }
-        if (dateMin) { dateMin.onmousewheel = gearMouseRolling }
-      }
     }
     // 公历农历选择
     function convertTap (type) {
@@ -301,7 +178,7 @@ class RuiDatepicker {
         let dateYY = valYY % passY + _self.minY
         let type = _self.type ? 0 : 1
         // 农历转公历前判断是否有闰月
-        let rmNum = LunarCal[valYY].Intercalation ? LunarCal[valYY].Intercalation : 0
+        let rmNum = calendar.LunarCal[valYY].Intercalation ? calendar.LunarCal[valYY].Intercalation : 0
         if (!_self.type && rmNum) {
           if (rmNum === (dateMM - 1)) {
             dateMM = -(dateMM - 1)
@@ -309,9 +186,9 @@ class RuiDatepicker {
             dateMM = dateMM - 1
           }
         }
-        let objDate = calendarConvert(type, dateYY, dateMM, dateDD)// 返回转换对象
+        let objDate = calendar.calendarConvert(type, dateYY, dateMM, dateDD)// 返回转换对象
         // 公历转农历后判断是否有闰月
-        let rmTip = LunarCal[objDate.yy - _self.minY].Intercalation ? LunarCal[objDate.yy - _self.minY].Intercalation : 0
+        let rmTip = calendar.LunarCal[objDate.yy - _self.minY].Intercalation ? calendar.LunarCal[objDate.yy - _self.minY].Intercalation : 0
         if (rmTip && _self.type) {
           if (objDate.mm < 0) {
             objDate.mm = -objDate.mm + 1
@@ -323,7 +200,7 @@ class RuiDatepicker {
         _self.gearDate.querySelector('.dateMM').setAttribute('val', objDate.mm - 1)
         _self.gearDate.querySelector('.dateDD').setAttribute('val', objDate.dd - 1)
         _self.gearDate.querySelector('.dateYY').setAttribute('top', '')
-        setDateGearTooth()
+        setDateGearTooth(true)
       }
     }
     // 初始化年月日插件默认值
@@ -371,8 +248,8 @@ class RuiDatepicker {
         let dateYY = dateArr.yy % passY + _self.minY
         let dateMM = dateArr.mm + 1
         let dateDD = dateArr.dd + 1
-        let objDate = calendarConvert(0, dateYY, dateMM, dateDD)// 返回转换对象
-        let isRm = LunarCal[objDate.yy - _self.minY].Intercalation ? LunarCal[objDate.yy - _self.minY].Intercalation : 0// 判断是不是闰年
+        let objDate = calendar.calendarConvert(0, dateYY, dateMM, dateDD)// 返回转换对象
+        let isRm = calendar.LunarCal[objDate.yy - _self.minY].Intercalation ? calendar.LunarCal[objDate.yy - _self.minY].Intercalation : 0// 判断是不是闰年
         if (objDate.mm < 0) {
           objDate.mm = -objDate.mm + 1// 返回的负数为 闰年
         } else {
@@ -393,7 +270,7 @@ class RuiDatepicker {
       setDateGearTooth()
     }
     // 重置日期节点个数+设置日期
-    function setDateGearTooth () {
+    function setDateGearTooth (isUpdate) {
       let passY = _self.maxY - _self.minY + 1
       let dateYY = _self.gearDate.querySelector('.dateYY')
       let yyVal = ''
@@ -406,22 +283,9 @@ class RuiDatepicker {
           itemStr += "<div class='tooth'>" + (_self.minY + p) + '</div>'
         }
         dateYY.innerHTML = itemStr
-        let top = Math.floor(parseFloat(dateYY.getAttribute('top')))
-        if (!isNaN(top)) {
-          if (top % 2 !== 0) { top = top + 1 }
-          top > 8 && (top = 8)
-          let minTop = 8 - (passY - 1) * 2
-          top < minTop && (top = minTop)
-          translateY(dateYY, top)
-          dateYY.setAttribute('top', top + 'em')
-          yyVal = Math.abs(top - 8) / 2
-          dateYY.setAttribute('val', yyVal)
-        } else {
-          translateY(dateYY, (8 - yyVal * 2))
-          dateYY.setAttribute('top', (8 - yyVal * 2) + 'em')
+        if (isUpdate) {
+          _self.yyScroll.update()
         }
-      } else {
-        // return;
       }
       let dateMM = _self.gearDate.querySelector('.dateMM')
       let maxM
@@ -431,7 +295,7 @@ class RuiDatepicker {
         // 得到月份的值
         mmVal = parseInt(dateMM.getAttribute('val'))
         // 判断否有闰月
-        let rmNum = LunarCal[yyVal].Intercalation ? LunarCal[yyVal].Intercalation : 0
+        let rmNum = calendar.LunarCal[yyVal].Intercalation ? calendar.LunarCal[yyVal].Intercalation : 0
         if (rmNum && _self.type) {
           maxM = 12
         } else {
@@ -445,7 +309,6 @@ class RuiDatepicker {
           } else {
             maxM = _self.maxM - 1
           }
-          // maxM = _self.maxM - 1;
         }
         // 当年份到达最小值
         if (yyVal === 0) {
@@ -454,7 +317,6 @@ class RuiDatepicker {
           } else {
             minM = _self.minM
           }
-          // minM = _self.minM - 1;
         }
         // p 当前节点前后需要展示的节点个数
         for (let p = 0; p < maxM - minM + 1; p++) {
@@ -463,11 +325,11 @@ class RuiDatepicker {
           if (_self.type) {
             // 闰月
             if (rmNum && rmNum === p) {
-              strVal = getChinese('rm', strVal - 1)
+              strVal = calendar.getChinese('rm', strVal - 1)
             } else if (rmNum && rmNum < p) {
-              strVal = getChinese('mm', strVal - 1)
+              strVal = calendar.getChinese('mm', strVal - 1)
             } else {
-              strVal = getChinese('mm', strVal)
+              strVal = calendar.getChinese('mm', strVal)
             }
           }
           itemStr += "<div class='tooth'>" + strVal + '</div>'
@@ -480,10 +342,10 @@ class RuiDatepicker {
           mmVal = maxM
           dateMM.setAttribute('val', mmVal)
         }
-        translateY(dateMM, (8 - (mmVal - minM) * 2))
-        dateMM.setAttribute('top', 8 - (mmVal - minM) * 2 + 'em')
-      } else {
-        // return;
+
+        if (isUpdate) {
+          _self.mmScroll.update()
+        }
       }
       let dateDD = _self.gearDate.querySelector('.dateDD')
       if (dateDD && dateDD.getAttribute('val')) {
@@ -491,8 +353,7 @@ class RuiDatepicker {
         // 得到日期的值
         let ddVal = parseInt(dateDD.getAttribute('val'))
         // 返回月份的天数
-        let maxMonthDays = calcDays(yyVal, mmVal)
-        // console.log(yyVal+','+mmVal+','+maxMonthDays);
+        let maxMonthDays = calendar.calcDays(_self.type, _self.minY, yyVal, mmVal)
         // p 当前节点前后需要展示的节点个数
         let maxD = maxMonthDays - 1
         let minD = 0
@@ -503,7 +364,6 @@ class RuiDatepicker {
           } else {
             maxD = _self.maxD - 1
           }
-          // maxD = _self.maxD - 1;
         }
         // 当年、月到达最小值
         if (yyVal === 0 && mmVal + 1 === 2) {
@@ -514,7 +374,7 @@ class RuiDatepicker {
           }
         }
         for (let p = 0; p < maxD - minD + 1; p++) {
-          let strVal = _self.type ? getChinese('dd', minD + p + 1) : (minD + p + 1)
+          let strVal = _self.type ? calendar.getChinese('dd', minD + p + 1) : (minD + p + 1)
           itemStr += "<div class='tooth'>" + strVal + '</div>'
         }
         dateDD.innerHTML = itemStr
@@ -525,574 +385,41 @@ class RuiDatepicker {
           ddVal = minD
           dateDD.setAttribute('val', ddVal)
         }
-        translateY(dateDD, (8 - (ddVal - minD) * 2))
-        dateDD.setAttribute('top', 8 - (ddVal - minD) * 2 + 'em')
-      } else {
-        // return;
+        if (isUpdate) {
+          _self.ddScroll.update()
+        }
       }
+
       let dateHH = _self.gearDate.querySelector('.dateHH')
       if (dateHH && dateHH.getAttribute('val')) {
-        let hhVal = parseInt(dateHH.getAttribute('val'))
-        // itemStr = "<div class='tooth'>未知</div>";
         itemStr = ''
         for (let p = 0; p < 24; p++) {
-          let strVal = _self.type ? getChinese('hh', p) : (p < 10 ? '0' + p : p)
+          let strVal = _self.type ? calendar.getChinese('hh', p) : (p < 10 ? '0' + p : p)
           itemStr += "<div class='tooth'>" + strVal + '时</div>'
         }
         dateHH.innerHTML = itemStr
-        translateY(dateHH, (8 - hhVal * 2))
-        dateHH.setAttribute('top', (8 - hhVal * 2) + 'em')
+        if (isUpdate) {
+          _self.hhScroll.update()
+        }
       }
 
       let dateMin = _self.gearDate.querySelector('.dateMin')
       if (dateMin && dateMin.getAttribute('val')) {
-        let minVal = parseInt(dateMin.getAttribute('val'))
-        // itemStr = "<div class='tooth'>未知</div>";
         itemStr = ''
         for (let p = 0; p < 60; p++) {
           let strVal = (p < 10 ? '0' + p : p)
           itemStr += "<div class='tooth'>" + strVal + '分</div>'
         }
         dateMin.innerHTML = itemStr
-        translateY(dateMin, (8 - minVal * 2))
-        dateMin.setAttribute('top', (8 - minVal * 2) + 'em')
+        if (isUpdate) {
+          _self.minScroll.update()
+        }
       }
       getCalendarDate()// 设置日期
-    }
-    function translateY (element, top) {
-      MIP.util.css(element, 'transform', 'translate(0,' + top + 'em)')
-      MIP.util.css(element, '-webkit-transform', 'translate(0,' + top + 'em)')
-      MIP.util.css(element, '-moz-transform', 'translate(0,' + top + 'em)')
-      MIP.util.css(element, '-ms-transform', 'translate(0,' + top + 'em)')
-      MIP.util.css(element, '-o-transform', 'translate(0,' + top + 'em)')
-    }
-    // 求月份最大天数
-    function calcDays (year, month) {
-      // 农历查询数据
-      if (_self.type === 1) {
-        if (LunarCal[year].MonthDays[month]) {
-          return 30
-        } else {
-          return 29
-        }
-      } else {
-        if (month === 1) {
-          year += _self.minY
-          if ((year % 4 === 0 && year % 100 !== 0) || (year % 400 === 0 && year % 4000 !== 0)) {
-            return 29
-          } else {
-            return 28
-          }
-        } else {
-          if (month === 3 || month === 5 || month === 8 || month === 10) {
-            return 30
-          } else {
-            return 31
-          }
-        }
-      }
-    }
-    // 中文转换,type:rm闰月，mm月份，dd日期，num值
-    function getChinese (type, num) {
-      let rmArr = ['闰正月', '闰二月', '闰三月', '闰四月', '闰五月', '闰六月', '闰七月', '闰八月', '闰九月', '闰十月', '闰冬月', '闰腊月']
-      let mmArr = ['正月', '二月', '三月', '四月', '五月', '六月', '七月', '八月', '九月', '十月', '十一月', '十二月']
-      let ddArr = ['初一', '初二', '初三', '初四', '初五', '初六', '初七', '初八', '初九', '初十', '十一', '十二', '十三', '十四', '十五', '十六', '十七', '十八', '十九', '二十', '廿一', '廿二', '廿三', '廿四', '廿五', '廿六', '廿七', '廿八', '廿九', '三十', '三十一']
-      let hhArr = ['0子', '1丑', '2丑', '3寅', '4寅', '5卯', '6卯', '7辰', '8辰', '9巳', '10巳', '11午', '12午', '13未', '14未', '15申', '16申', '17酉', '18酉', '19戌', '20戌', '21亥', '22亥', '23子']
-      if (type === 'rm') return rmArr[num - 1]
-      if (type === 'mm') return mmArr[num - 1]
-      if (type === 'dd') return ddArr[num - 1]
-      if (type === 'hh') return hhArr[num]
-    }
-
-    /**
-     *
-     * @param {string} M =0公历到农历,M=1农历到公历
-     * @param {string} dateYY 年
-     * @param {string} dateMM 月，闰月为 负数
-     * @param {string} dateDD 日
-     */
-    function calendarConvert (M, dateYY, dateMM, dateDD) {
-      let SolarYear = parseInt(dateYY)
-      let SolarMonth = parseInt(dateMM)
-      let SolarDate = parseInt(dateDD)
-      let LunarYear = parseInt(dateYY)
-      let LunarMonth = parseInt(dateMM)
-      let LunarDate = parseInt(dateDD)
-
-      let FIRSTYEAR = 1940// 最小年份
-      // let LASTYEAR = 2030// 最大年份
-
-      // 西曆年每月之日數
-      // let SolarCal = [31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31]
-
-      // 西曆年每月之累積日數, 平年與閏年
-      let SolarDays = [0, 31, 59, 90, 120, 151, 181, 212, 243, 273, 304, 334, 365, 396, 0, 31, 60, 91, 121, 152, 182, 213, 244, 274, 305, 335, 366, 397]
-      let i
-      // 公历转农历
-      if (M === 0) {
-        let sm = SolarMonth - 1
-
-        let leap = GetLeap(SolarYear)
-
-        // let d = (sm === 1) ? leap + 28 : SolarCal[sm]
-
-        let y = SolarYear - FIRSTYEAR
-        let acc = SolarDays[leap * 14 + sm] + SolarDate
-        let kc = acc + LunarCal[y].BaseKanChih
-        let Age = kc % 60
-        Age = (Age < 22) ? 22 - Age : 82 - Age
-
-        Age = Age + 3
-        if (Age < 10) Age = Age + 60
-
-        if (acc <= LunarCal[y].BaseDays) {
-          y--
-          LunarYear = SolarYear - 1
-          leap = GetLeap(LunarYear)
-          sm += 12
-          acc = SolarDays[leap * 14 + sm] + SolarDate
-        } else {
-          LunarYear = SolarYear
-        }
-        let l1 = LunarCal[y].BaseDays
-        for (i = 0; i < 13; i++) {
-          let l2 = l1 + LunarCal[y].MonthDays[i] + 29
-          if (acc <= l2) break
-          l1 = l2
-        }
-        let LunarMonth = i + 1
-        let LunarDate = acc - l1
-        let im = LunarCal[y].Intercalation
-        if (im !== 0 && LunarMonth > im) {
-          LunarMonth--
-          if (LunarMonth === im) LunarMonth = -im
-        }
-        if (LunarMonth > 12) LunarMonth -= 12
-
-        return {yy: LunarYear, mm: LunarMonth, dd: LunarDate}
-      } else {
-        /* 农历转公历 */
-        let y = LunarYear - FIRSTYEAR
-        let im = LunarCal[y].Intercalation
-        let lm = LunarMonth
-
-        if (im !== 0) {
-          if (lm > im) {
-            lm++
-          } else if (lm === -im) {
-            lm = im + 1
-          }
-        }
-        lm--
-
-        let acc = 0
-        for (i = 0; i < lm; i++) {
-          acc += LunarCal[y].MonthDays[i] + 29
-        }
-
-        acc += LunarCal[y].BaseDays + LunarDate
-
-        let leap = GetLeap(LunarYear)
-
-        for (i = 13; i >= 0; i--) {
-          if (acc > SolarDays[leap * 14 + i]) break
-        }
-        SolarDate = acc - SolarDays[leap * 14 + i]
-        if (i <= 11) {
-          SolarYear = LunarYear
-          SolarMonth = i + 1
-        } else {
-          SolarYear = LunarYear + 1
-          SolarMonth = i - 11
-        }
-        // return SolarYear + "-" + SolarMonth + "-" + SolarDate;
-        return {yy: SolarYear, mm: SolarMonth, dd: SolarDate}
-      }
-    }
-    /* 闰年, 返回 0 平年, 1 闰年 */
-    function GetLeap (year) {
-      if (year % 400 === 0) { return 1 } else if (year % 100 === 0) { return 0 } else if (year % 4 === 0) { return 1 } else { return 0 }
-    }
-    // 农历对象数据处理
-    function TagLunarCal (d, i, w, k, m1, m2, m3, m4, m5, m6, m7, m8, m9, m10, m11, m12, m13) {
-      this.BaseDays = d /* 1 月 1 日到正月初一的累计日 */
-      this.Intercalation = i /* 闰月月份. 0==此年沒有闰月 */
-      this.BaseWeekday = w /* 此年 1 月 1 日为星期减 1 */
-      this.BaseKanChih = k /* 此年 1 月 1 日之干支序号减 1 */
-      this.MonthDays = [m1, m2, m3, m4, m5, m6, m7, m8, m9, m10, m11, m12, m13] /* 此农历年每月之大小, 0==小月(29日), 1==大月(30日) */
     }
     // 补零
     function addZero (i) {
       return i < 10 ? '0' + i : '' + i
-    }
-    // PC滑动鼠标滚轮处理
-    function gearMouseRolling (e) {
-      e = e || window.event
-      let dir = true
-      if (e.wheelDelta) {
-        dir = e.wheelDelta > 0 // ie和chrome
-      } else {
-        dir = e.detail < 0// firefox
-      }
-      // dir:true向上滚动，false向下滚动
-      let moveDir = dir ? 21 : -21
-      // 移动前
-      e.preventDefault()
-      let target = e.target
-      while (true) {
-        if (!target.classList.contains('gear')) {
-          target = target.parentElement
-        } else {
-          break
-        }
-      }
-      clearInterval(target['int_' + target.id])
-      target['old_' + target.id] = 0
-      target['o_t_' + target.id] = (new Date()).getTime()
-      let top = target.getAttribute('top')
-      if (top) {
-        target['o_d_' + target.id] = parseFloat(top.replace(/em/g, ''))
-      } else {
-        target['o_d_' + target.id] = 0
-      }
-      // 移动中
-      target['new_' + target.id] = moveDir
-      target['n_t_' + target.id] = (new Date()).getTime() + 360
-      let f = (target['new_' + target.id] - target['old_' + target.id]) * 18 / 370
-      target['pos_' + target.id] = target['o_d_' + target.id] + f
-      target.setAttribute('top', target['pos_' + target.id] + 'em')
-      // 移动后
-      let flag = (target['new_' + target.id] - target['old_' + target.id]) / (target['n_t_' + target.id] - target['o_t_' + target.id])
-      if (Math.abs(flag) <= 0.2) {
-        target['spd_' + target.id] = (flag < 0 ? -0.08 : 0.08)
-      } else {
-        if (Math.abs(flag) <= 0.5) {
-          target['spd_' + target.id] = (flag < 0 ? -0.16 : 0.16)
-        } else {
-          target['spd_' + target.id] = flag / 2
-        }
-      }
-      if (!target['pos_' + target.id]) {
-        target['pos_' + target.id] = 0
-      }
-      if (e.preventDefault) { e.preventDefault() }
-      rollGear(target)
-      return false
-    }
-    // PC滑动鼠标处理
-    function gearMouseSlide (e) {
-      e.preventDefault()
-      let target = e.target
-      let targetPc = target
-      let mousedownTip = false// 鼠标滑动锁
-      while (true) {
-        if (!target.classList.contains('gear')) {
-          target = target.parentElement
-        } else {
-          break
-        }
-      }
-      clearInterval(target['int_' + target.id])
-      target['old_' + target.id] = e.screenY
-      target['o_t_' + target.id] = (new Date()).getTime()
-      let top = target.getAttribute('top')
-      if (top) {
-        target['o_d_' + target.id] = parseFloat(top.replace(/em/g, ''))
-      } else {
-        target['o_d_' + target.id] = 0
-      }
-      document.onmousemove = function (e) {
-        mousedownTip = true// 鼠标滑动锁
-        e = e || window.event
-        e.preventDefault()
-        let target = targetPc// 鼠标兼容处理
-        while (true) {
-          if (!target) break
-          if (!target.classList.contains('gear')) {
-            target = target.parentElement
-          } else {
-            break
-          }
-        }
-        if (!target) return
-        target['new_' + target.id] = e.screenY
-        target['n_t_' + target.id] = (new Date()).getTime()
-        let f = (target['new_' + target.id] - target['old_' + target.id]) * 18 / 370
-        target['pos_' + target.id] = target['o_d_' + target.id] + f
-        translateY(target, target['pos_' + target.id])
-        target.setAttribute('top', target['pos_' + target.id] + 'em')
-      }
-      // 释放鼠标
-      document.onmouseup = function (e) {
-        // 阻止点击
-        if (!mousedownTip) {
-          document.onmousemove = null
-          document.onmouseup = null
-          return false
-        }
-        e = e || window.event
-        e.preventDefault()
-        let target = targetPc// 鼠标兼容处理
-        while (true) {
-          if (!target) break
-          if (!target.classList.contains('gear')) {
-            target = target.parentElement
-          } else {
-            break
-          }
-        }
-        if (!target) return
-        let flag = (target['new_' + target.id] - target['old_' + target.id]) / (target['n_t_' + target.id] - target['o_t_' + target.id])
-        if (Math.abs(flag) <= 0.2) {
-          target['spd_' + target.id] = (flag < 0 ? -0.08 : 0.08)
-        } else {
-          if (Math.abs(flag) <= 0.5) {
-            target['spd_' + target.id] = (flag < 0 ? -0.16 : 0.16)
-          } else {
-            target['spd_' + target.id] = flag / 2
-          }
-        }
-        if (!target['pos_' + target.id]) {
-          target['pos_' + target.id] = 0
-        }
-        rollGear(target)
-        document.onmousemove = null
-        document.onmouseup = null
-      }
-    }
-    // 触摸开始
-    function gearTouchStart (e) {
-      e.preventDefault()
-      let target = e.target
-      target['touchTip'] = false// 滑动锁
-      while (true) {
-        if (!target.classList.contains('gear')) {
-          target = target.parentElement
-        } else {
-          break
-        }
-      }
-      clearInterval(target['int_' + target.id])
-      target['old_' + target.id] = e.targetTouches[0].screenY
-      target['o_t_' + target.id] = (new Date()).getTime()
-      let top = target.getAttribute('top')
-      if (top) {
-        target['o_d_' + target.id] = parseFloat(top.replace(/em/g, ''))
-      } else {
-        target['o_d_' + target.id] = 0
-      }
-    }
-    // 手指移动
-    function gearTouchMove (e) {
-      e.preventDefault()
-      let target = e.target
-      target['touchTip'] = true// 滑动锁
-      while (true) {
-        if (!target.classList.contains('gear')) {
-          target = target.parentElement
-        } else {
-          break
-        }
-      }
-      target['new_' + target.id] = e.targetTouches[0].screenY
-      target['n_t_' + target.id] = (new Date()).getTime()
-      let f = (target['new_' + target.id] - target['old_' + target.id]) * 18 / 370
-      target['pos_' + target.id] = target['o_d_' + target.id] + f
-      translateY(target, target['pos_' + target.id])
-      target.setAttribute('top', target['pos_' + target.id] + 'em')
-    }
-    // 离开屏幕
-    function gearTouchEnd (e) {
-      e.preventDefault()
-      let target = e.target
-      if (!target['touchTip']) return false
-      while (true) {
-        if (!target.classList.contains('gear')) {
-          target = target.parentElement
-        } else {
-          break
-        }
-      }
-      let flag = (target['new_' + target.id] - target['old_' + target.id]) / (target['n_t_' + target.id] - target['o_t_' + target.id])
-      if (Math.abs(flag) <= 0.2) {
-        target['spd_' + target.id] = (flag < 0 ? -0.08 : 0.08)
-      } else {
-        if (Math.abs(flag) <= 0.5) {
-          target['spd_' + target.id] = (flag < 0 ? -0.16 : 0.16)
-        } else {
-          target['spd_' + target.id] = flag / 2
-        }
-      }
-      if (!target['pos_' + target.id]) {
-        target['pos_' + target.id] = 0
-      }
-      rollGear(target)
-    }
-    // 缓动效果
-    function rollGear (target) {
-      let d = 0
-      let stopGear = false
-      let passY = _self.maxY - _self.minY + 1
-      clearInterval(target['int_' + target.id])
-      target['int_' + target.id] = setInterval(function () {
-        let pos = target['pos_' + target.id]
-        let speed = target['spd_' + target.id] * Math.exp(-0.03 * d)
-        pos += speed
-        if (Math.abs(speed) > 0.1) {} else {
-          speed = 0.1
-          let b = Math.round(pos / 2) * 2
-          if (Math.abs(pos - b) < 0.02) {
-            stopGear = true
-          } else {
-            if (pos > b) {
-              pos -= speed
-            } else {
-              pos += speed
-            }
-          }
-        }
-        if (pos > 8) {
-          pos = 8
-          stopGear = true
-        }
-
-        // let minTop
-        let dateYY
-        let dateMM
-        let yyVal
-        let mmVal
-        switch (target.getAttribute('data-datetype')) {
-          case 'dateYY':
-            let minTop = 8 - (passY - 1) * 2
-            if (pos < minTop) {
-              pos = minTop
-              stopGear = true
-            }
-            if (stopGear) {
-              let gearVal = Math.abs(pos - 8) / 2
-              setGear(target, gearVal)
-              clearInterval(target['int_' + target.id])
-            }
-            break
-          case 'dateMM':
-            dateYY = _self.gearDate.querySelector('.dateYY')
-            // 得到年份的值
-            yyVal = parseInt(dateYY.getAttribute('val'))
-            // 判断否有闰月
-            let rmNum = LunarCal[yyVal].Intercalation ? LunarCal[yyVal].Intercalation : 0
-            let maxM
-            if (rmNum && _self.type) {
-              maxM = 12
-            } else {
-              maxM = 11
-            }
-            let minM = 0
-            // 当年份到达最大值
-            if (yyVal === passY - 1) {
-              if (_self.type) {
-                maxM = _self.maxM - 1
-              } else {
-                maxM = _self.maxM - 1
-              }
-              // maxM = _self.maxM - 1;
-            }
-            // 当年份到达最小值
-            if (yyVal === 0) {
-              if (_self.type) {
-                minM = _self.minM - 1
-              } else {
-                minM = _self.minM
-              }
-              // minM = _self.minM - 1;
-            }
-            minTop = 8 - (maxM - minM) * 2
-            if (pos < minTop) {
-              pos = minTop
-              stopGear = true
-            }
-            if (stopGear) {
-              let gearVal = Math.abs(pos - 8) / 2 + minM
-              setGear(target, gearVal)
-              clearInterval(target['int_' + target.id])
-            }
-            break
-          case 'dateDD':
-            dateYY = _self.gearDate.querySelector('.dateYY')
-            dateMM = _self.gearDate.querySelector('.dateMM')
-            // 得到年份的值
-            yyVal = parseInt(dateYY.getAttribute('val'))
-            // 得到月份的值
-            mmVal = parseInt(dateMM.getAttribute('val'))
-            // 返回月份的天数
-            let maxMonthDays = calcDays(yyVal, mmVal)
-            let maxD = maxMonthDays - 1
-            let minD = 0
-            // 当年份月份到达最大值
-            if (yyVal === passY - 1 && mmVal + 1 === 12) {
-              if (_self.type) {
-                maxD = _self.maxD - 15
-              } else {
-                maxD = _self.maxD - 1
-              }
-              // maxD = _self.maxD - 1;
-            }
-            // 当年、月到达最小值
-            if (yyVal === 0 && mmVal + 1 === 2) {
-              if (_self.type) {
-                minD = _self.minD - 1
-              } else {
-                minD = _self.minD + 6
-              }
-              // minD = _self.minD - 1;
-            }
-            minTop = 8 - (maxD - minD) * 2
-            if (pos < minTop) {
-              pos = minTop
-              stopGear = true
-            }
-            if (stopGear) {
-              let gearVal = Math.abs(pos - 8) / 2 + minD
-              setGear(target, gearVal)
-              clearInterval(target['int_' + target.id])
-            }
-            break
-          case 'dateHH':
-            minTop = 8 - (24 - 1) * 2
-            if (pos < minTop) {
-              pos = minTop
-              stopGear = true
-            }
-            if (stopGear) {
-              let gearVal = Math.abs(pos - 8) / 2
-              setGear(target, gearVal)
-              clearInterval(target['int_' + target.id])
-            }
-            break
-          case 'dateMin':
-            minTop = 8 - (60 - 1) * 2
-            if (pos < minTop) {
-              pos = minTop
-              stopGear = true
-            }
-            if (stopGear) {
-              let gearVal = Math.abs(pos - 8) / 2
-              setGear(target, gearVal)
-              clearInterval(target['int_' + target.id])
-            }
-            break
-          default:
-        }
-        target['pos_' + target.id] = pos
-        translateY(target, pos)
-        target.setAttribute('top', pos + 'em')
-        d++
-      }, 6)
-    }
-    // 控制插件滚动后停留的值
-    function setGear (target, val) {
-      val = Math.round(val)
-      target.setAttribute('val', val)
-      setDateGearTooth()
     }
     // 取消 type 判断是不是点击完成后调用
     function closeMobileCalendar (e, type) {
@@ -1144,14 +471,14 @@ class RuiDatepicker {
             gongliHourStr = '时辰未知'
           } else {
             gongliHourStr = d.hh + '时'
-            nongliHourStr = getChinese('hh', d.hh) + '时'
+            nongliHourStr = calendar.getChinese('hh', d.hh) + '时'
           }
         }
         // 设置公历选择时间
         confirmGongli.innerHTML = d.yy + '年' + d.mm + '月' + d.dd + '日 ' + gongliHourStr
         // 设置农历选择时间
-        let mmChina = d._mm < 0 ? getChinese('rm', -d._mm) : getChinese('mm', d._mm)
-        confirmNongli.innerHTML = d._yy + '年' + mmChina + '' + getChinese('dd', d._dd) + ' ' + nongliHourStr
+        let mmChina = d._mm < 0 ? calendar.getChinese('rm', -d._mm) : calendar.getChinese('mm', d._mm)
+        confirmNongli.innerHTML = d._yy + '年' + mmChina + '' + calendar.getChinese('dd', d._dd) + ' ' + nongliHourStr
         return false
       }
       let dateId = _self.trigger.getAttribute('data-toid-date')
@@ -1185,11 +512,11 @@ class RuiDatepicker {
       if (dateId) {
         let hourStr = ''
         if (_self.type) {
-          let mmChina = d._mm < 0 ? getChinese('rm', -d._mm) : getChinese('mm', d._mm)
+          let mmChina = d._mm < 0 ? calendar.getChinese('rm', -d._mm) : calendar.getChinese('mm', d._mm)
           if (hourId) {
-            hourStr = getChinese('hh', d.hh) + '时'
+            hourStr = calendar.getChinese('hh', d.hh) + '时'
           }
-          _self.trigger.value = d._yy + '年' + mmChina + '' + getChinese('dd', d._dd) + ' ' + hourStr
+          _self.trigger.value = d._yy + '年' + mmChina + '' + calendar.getChinese('dd', d._dd) + ' ' + hourStr
         } else {
           if (hourId) {
             hourStr = d.hh + '时'
@@ -1230,7 +557,7 @@ class RuiDatepicker {
           dateHH = parseInt(Math.round(_self.gearDate.querySelector('.dateHH').getAttribute('val')))
         }
         // 判断否有闰月
-        let rmNum = LunarCal[valYY].Intercalation ? LunarCal[valYY].Intercalation : 0
+        let rmNum = calendar.LunarCal[valYY].Intercalation ? calendar.LunarCal[valYY].Intercalation : 0
         if (_self.type && rmNum) {
           if (rmNum === (dateMM - 1)) {
             dateMM = -(dateMM - 1)
@@ -1238,20 +565,20 @@ class RuiDatepicker {
             dateMM = dateMM - 1
           }
         }
-        let objDate = calendarConvert(_self.type, dateYY, dateMM, dateDD)
+        let objDate = calendar.calendarConvert(_self.type, dateYY, dateMM, dateDD)
         let info = _self.gearDate.querySelector('.lcalendar_info')
         if (_self.type) {
           _self.trigger.setAttribute('data-type', 1)
-          let mmChina = dateMM < 0 ? getChinese('rm', -dateMM) : getChinese('mm', dateMM)
+          let mmChina = dateMM < 0 ? calendar.getChinese('rm', -dateMM) : calendar.getChinese('mm', dateMM)
           let hhStr = ''
           if (hourOn) {
             if (dateHH < 0) {
               hhStr = '时辰未知'
             } else {
-              hhStr = getChinese('hh', dateHH) + '时'
+              hhStr = calendar.getChinese('hh', dateHH) + '时'
             }
           }
-          info.innerHTML = dateYY + '年' + mmChina + '' + getChinese('dd', dateDD) + ' ' + hhStr
+          info.innerHTML = dateYY + '年' + mmChina + '' + calendar.getChinese('dd', dateDD) + ' ' + hhStr
           return {
             yy: objDate.yy,
             mm: objDate.mm,
