+++ sites/m.dishen.com/components/mip-dishen-datepicker/style/index.less
@@ -151,7 +151,7 @@
     position: absolute;
     z-index: 9902;
     width: 5.5em;
-    margin-top: -6em;
+    margin-top: 2em;
     cursor: pointer;
   }
